+++ CHANGES.md
@@ -151,6 +151,7 @@
     -   Fix bold regexp issue([GH-325][])
     -   Fix italic and punctual character issue([GH-359][])
     -   Fix table align issue when column contains escaped bar([GH-308][])
+    -   Fix nested block fill-paragraph issue([GH-366][])
 
   [gh-171]: https://github.com/jrblevin/markdown-mode/issues/171
   [gh-216]: https://github.com/jrblevin/markdown-mode/issues/216
@@ -207,6 +208,7 @@
   [gh-349]: https://github.com/jrblevin/markdown-mode/issues/349
   [gh-350]: https://github.com/jrblevin/markdown-mode/pull/350
   [gh-359]: https://github.com/jrblevin/markdown-mode/issues/359
+  [gh-366]: https://github.com/jrblevin/markdown-mode/issues/366
   [gh-369]: https://github.com/jrblevin/markdown-mode/pull/369
   [gh-378]: https://github.com/jrblevin/markdown-mode/pull/378
   [gh-383]: https://github.com/jrblevin/markdown-mode/issues/383